+++ osmos-core/src/sensor.rs
@@ -1,11 +1,11 @@
 pub struct Sensor {
-    pub range: f64,
+    pub range: f32,
     // [up, right, down, left]
-    pub data_list: [f64; 4],
+    pub data_list: [f32; 4],
 }
 
 impl Sensor {
-    pub fn new(range: f64) -> Self {
+    pub fn new(range: f32) -> Self {
         Self {
             range,
             data_list: [0.0; 4],
